# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "The Nexus API",
    "version": "1.0.0",
    "description": "API documentation for The Nexus platform, standardized with v1 routes."
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/api/v1/introductions/{id}/accept": {
      "post": {
        "summary": "Accept an introduction request",
        "tags": [
          "Introductions",
          "Professionals"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Introduction request ID",
            "example": "intro_123"
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "message": {
                    "type": "string",
                    "description": "Optional response message to the HR partner",
                    "maxLength": 500,
                    "example": "I'm interested in this opportunity and would like to discuss further."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Introduction request accepted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Introduction request accepted successfully"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "introduction": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "string"
                            },
                            "sentToProfessionalId": {
                              "type": "string"
                            },
                            "sentById": {
                              "type": "string"
                            },
                            "jobRoleId": {
                              "type": "string"
                            },
                            "status": {
                              "type": "string",
                              "enum": [
                                "ACCEPTED"
                              ]
                            },
                            "message": {
                              "type": "string"
                            },
                            "professionalResponse": {
                              "type": "string",
                              "nullable": true
                            },
                            "sentAt": {
                              "type": "string",
                              "format": "date-time"
                            },
                            "responseDate": {
                              "type": "string",
                              "format": "date-time"
                            },
                            "expiresAt": {
                              "type": "string",
                              "format": "date-time"
                            },
                            "viewedByProfessional": {
                              "type": "boolean"
                            },
                            "viewedAt": {
                              "type": "string",
                              "format": "date-time"
                            },
                            "jobRole": {
                              "type": "object",
                              "properties": {
                                "id": {
                                  "type": "string"
                                },
                                "roleTitle": {
                                  "type": "string"
                                },
                                "roleDescription": {
                                  "type": "string"
                                },
                                "seniorityLevel": {
                                  "type": "string"
                                },
                                "locationCity": {
                                  "type": "string",
                                  "nullable": true
                                },
                                "locationState": {
                                  "type": "string",
                                  "nullable": true
                                },
                                "salaryRangeMin": {
                                  "type": "number",
                                  "nullable": true
                                },
                                "salaryRangeMax": {
                                  "type": "number",
                                  "nullable": true
                                },
                                "remoteOption": {
                                  "type": "string"
                                },
                                "employmentType": {
                                  "type": "string"
                                },
                                "responsibilities": {
                                  "type": "string",
                                  "nullable": true
                                },
                                "requirements": {
                                  "type": "string",
                                  "nullable": true
                                },
                                "preferredQualifications": {
                                  "type": "string",
                                  "nullable": true
                                },
                                "benefits": {
                                  "type": "string",
                                  "nullable": true
                                },
                                "isConfidential": {
                                  "type": "boolean"
                                },
                                "confidentialReason": {
                                  "type": "string",
                                  "nullable": true
                                }
                              }
                            },
                            "company": {
                              "type": "object",
                              "properties": {
                                "id": {
                                  "type": "string"
                                },
                                "companyName": {
                                  "type": "string"
                                },
                                "companyLogoUrl": {
                                  "type": "string",
                                  "nullable": true
                                },
                                "industry": {
                                  "type": "string"
                                },
                                "companySize": {
                                  "type": "string"
                                },
                                "headquartersLocation": {
                                  "type": "string"
                                },
                                "companyWebsite": {
                                  "type": "string",
                                  "nullable": true
                                },
                                "companyDescription": {
                                  "type": "string"
                                }
                              }
                            },
                            "sentBy": {
                              "type": "object",
                              "properties": {
                                "id": {
                                  "type": "string"
                                },
                                "firstName": {
                                  "type": "string"
                                },
                                "lastName": {
                                  "type": "string"
                                },
                                "jobTitle": {
                                  "type": "string"
                                },
                                "profilePhotoUrl": {
                                  "type": "string",
                                  "nullable": true
                                },
                                "linkedinUrl": {
                                  "type": "string",
                                  "nullable": true
                                }
                              }
                            }
                          }
                        },
                        "contactDetailsUnlocked": {
                          "type": "boolean",
                          "description": "Whether contact details are now available",
                          "example": true
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or introduction not actionable",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Introduction request not found or already responded to"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Unauthorized"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - professional access required",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Forbidden: Professional access required"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Introduction request not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Introduction request not found or already responded to"
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Validation failed"
                    },
                    "details": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "field": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Failed to accept introduction request"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}